Add basic JSON framework and stub tests
`runtype_options.json` contains the various parameters that are passed to
`run.sh`. Note that in this case, all of the options happen to require 11
parameters, but this is not required, as shown by `dialogue.py`.

`dialogue.py` currently doesn't have much of a dialogue at all, but later
commits are going to start loading configurations from another json file
(`testconfig.json` or something similar) and calling `execute_run` on the
dictionary `load`ed from that.

Lastly, `test.py` contains the current contents of `goMLQ.sh`. It creates
a dictionary for each run and "executes" it. Currently, it simply echoes
the `run.sh` invocation. There is also a silly run that should fail.