Clean up various silliness in rough commit
This is a rebase of `mab-cleanup`. See that branch if you want to see
the incremental changes.

Remove obsolete functions from dialogue.py

All of these removals are from the FIRST BLOOD^(TM) edition,
where I was still feeling out a few things. Benjamin Spriggs
implemented a far better version, so it's now unneeded.

---

Add whitespace to printing functions

I'm trying to avoid enormous walls of text here.

---

Add YES_CONSTS and EXIT_CONSTS

`foo in ['q', 'quit', 'exit']` gets a little old after a while.

---

Preserve parameter order in add_runtype

The parameter order is important, and we want to keep it consistent.
The current model contains a `set`, which does not preserve order.

An alternative is to sort the set by alphabetical order, or use
the `collections` library to implement an `OrderedSet` as a special
case of an `OrderedDict`. There is also a library on PyPi that also
implements this functionality.

I don't like changing the parameter
order, and it seems a little silly to add another dependency for
this. Using a list to track insertion order is pretty simple.