Preserve parameter order in add_runtype
The parameter order is important, and we want to keep it consistent.
The current model contains a `set`, which does not preserve order.

An alternative is to sort the set by alphabetical order, or use
the `collections` library to implement an `OrderedSet` as a special
case of an `OrderedDict`. There is also a library on PyPi that also
implements this functionality.

I don't like changing the parameter
order, and it seems a little silly to add another dependency for
this. Using a list to track insertion order is pretty simple.