Change runtype_options
This goes with the meeting that we held today.

Each runtype will now contain the following values:

* args, containing an ordered list of arguments. This is for the
purpose of dialogue and presenting each parameter to the user.

* annotations, containing key-value pairs of arguments and
descriptions for each one. This is for explaining each parameter
to the user.

* types, containing key-value pairs of arguments and types. This
is for defining valid user input for each argument.

* translations, containing key-value pairs of arguments and
.props values. Upon running in main, the program will take a Run
and its corresponding RunType and write values to its temporary
.props file.

* props_defaults, containing values that will be written by
default to the .props file without any user input at all.
These are intrinsic to the RunType.

For example, let's say that we have an RunType with the following:

    "SPAM" : {
      args : [
        "foo"
        "bar"
      ],

      annotations : {
        "foo" : "Lorem ipsum dolor sit amet, consectetur...",
        "bar" : "Sed blandit metus quis nulla dapibus..."
      },

      "types" : {
        "foo" : "string",
        "bar" : "integer"
      },

      "translations" : {
        "foo" : "controller.quux.spam.eggs.foo"
      }

      "props_defaults" : {
        "controller.muffin.scone" : false,
        "controller.waffle.pancake" : 50,
        "controller.bacon.sausage" : "eggs"
      }
    }

Upon a user creating a SPAM Run, the Dialogue will ask the user to
provide values for `foo` and `bar`, demanding a string and integer
as input, respectively. If needed, the user can obtain the annotations
(Lorem ipsum dolor sit amet...) to explain what he is supposed to input.
Let's say that he inputs `oatmeal` and `5`, respectively.

Upon running the above run, `main` will write the following to
the temporary .props file:

    controller.quux.spam.eggs.foo=oatmeal
    controller.muffin.scone=false
    controller.waffle.pancake=50
    controller.bacon.sausage=eggs

The remaining `bar` value will possibly be used in the actual JVM
invocation.